+++ minitalk_bonus.h
@@ -6,7 +6,7 @@
 /*   By: cjimenez <cjimenez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/01 17:07:31 by cjimenez          #+#    #+#             */
-/*   Updated: 2022/03/14 11:12:59 by cjimenez         ###   ########.fr       */
+/*   Updated: 2022/03/23 10:00:30 by cjimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdlib.h>
 # include <sys/signal.h>
 
-int	ft_isdigit(int c);
-int	ft_atoi(const char *str);
+int		ft_isdigit(int c);
+int64_t	ft_atoi(const char *str);
 
 #endif
